+++ src/layers/TextLayer.tsx
@@ -14,6 +14,7 @@ import calculateOffset from '../utils/calculateOffset';
 import unwrap from '../utils/unwrap';
 import wrap from '../utils/wrap';
 import PdfJs from '../vendors/PdfJs';
+import { TextLayerRenderEvent } from '../Viewer';
 import './textLayer.less';
 import WithScale from './WithScale';
 
@@ -25,9 +26,10 @@ interface TextLayerProps {
     rotation: number;
     scale: number;
     onJumpToMatch(pageIndex: number, top: number, left: number): void;
+    onTextLayerRender(e: TextLayerRenderEvent): void;
 }
 
-const TextLayer: React.FC<TextLayerProps> = ({ keywordRegexp, match, page, pageIndex, rotation, scale, onJumpToMatch }) => {
+const TextLayer: React.FC<TextLayerProps> = ({ keywordRegexp, match, page, pageIndex, rotation, scale, onJumpToMatch, onTextLayerRender }) => {
     const theme = useContext(ThemeContext);
     const containerRef = createRef<HTMLDivElement>();
     const renderTask = useRef<PdfJs.PageRenderTask>();
@@ -126,6 +128,9 @@ const TextLayer: React.FC<TextLayerProps> = ({ keywordRegexp, match, page, pageI
                         }
                     }
                     scrollToMatch();
+                    onTextLayerRender({
+                        ele: containerEle,
+                    });
                 },
                 () => {/**/},
             );